branch update
git-svn-id: https://svn.r-project.org/R/trunk@26734 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/library/ctest/R/cor.test.R
@@ -116,7 +116,7 @@ function(x, y, alternative = c("two.sided", "less", "greater"),
                     if(n <= 1290) # n*(n^2 - 1) does not overflow
                         .C("prho",
                            as.integer(n),
-                           as.integer(q + 1),
+                           as.double(q + 1),
                            p = double(1),
                            integer(1),
                            as.logical(lower.tail),
